Set alert appearance only once
An alert’s appearance is defined the moment it is shown. Though it can be reconfigured after it is shown, none of these changes will be visible, even if the alert was hidden (due to some other alert showing). That means that changing the alertViewStyle after the alert received the show message will not add or remove text fields, for example. The same applies to buttons. The only thing that can change is title and message labels.