Move button management to SDCAlertViewContentView
Since the alert’s content view displays buttons and needs to keep track of the configuration of the content, moving all button handling to SDCAlertViewContentView makes more sense than doing it in SDCAlertView itself. The public API for SDCAlertView still allows the user to make changes, but the implementation of these methods are simply passed to SDCAlertViewContentView.