Define text field placeholders earlier
Previously, the text fields’ placeholders were defined right before in a call tree started by -[SDCAlertView show]. This meant that placeholders set by the user were going to be overridden as soon as the alert was displayed. This problem can be avoided by defining the placeholders as early as possible.