Automatically update constraints when showing alert
An explicit call to setNeedsUpdateConstraints was required after an alert was added to the view hierarchy, so that the alert knew what size it could take (should it need to expand due to large content). That call is now made automatically when the view is added to the view hierarchy, which makes it easier for custom SDCAlertViewTransitioning objects to be implemented.