Show correct button titles in two-button alert
In an alert configuration with two button other button titles, but no cancel button, the first one of the two other buttons was shown twice. This commit fixes that and shows the first button on the left, the second button as the suggested button on the right (following UIAlertView standards).