Fix #102—Avoid retain cycle between SDCAlertView and SDCAlertController
The shouldDismissBlock and didDismissBlock of SDCAlertView had a reference to the SDCAlertController instance that was used to instantiate the legacy alert view. Since the alert controller also had a strong reference to the legacy alert view, this caused a retain cycle. The fix is to remove the reference to the alert controller in the aforementioned blocks and copy any values needed from the controller and use those in the blocks instead.