Allow for horizontal scrolling in buttons
The pan gesture that selects a new button when the user moves from one button to
another interfered with the scroll view's pan gesture. This change gives
priority to the scroll view and only if there is no scrolling will the button
selection gesture be handled.

It also un-selects a button if the user's finger is completely off the
collection view.

The tap gesture is also removed in favor of the collection view's native
selection handling.